fix : pantheon scale WIP

+++ gui/include/graphics/raylib/RaylibModels.hpp
@@ -41,7 +41,7 @@ namespace Zappy {
             // we try to assign an animation for the pantheon even if it is not referenced
             if (indexPantheon == -1) {
                 std::vector<int> allIndexes = {indexIdle, indexWalk, indexSleep, indexCry };
-                for (int i = 0; i < allIndexes.size(); i++) {
+                for (size_t i = 0; i < allIndexes.size(); i++) {
                     if (allIndexes[i] != -1) {
                         indexPantheon = allIndexes[i];
                         break;
